chore(suite-desktop): move thunks to separate file
