+++ packages/suite/src/components/suite/NavigationBar/components/DeviceSelector/index.tsx
@@ -1,4 +1,5 @@
 import { useState, useEffect, useRef } from 'react';
+
 import styled, { css } from 'styled-components';
 import { transparentize } from 'polished';
 
@@ -11,7 +12,7 @@ import { WalletLabeling } from 'src/components/suite';
 import { TrezorDevice } from 'src/types/suite';
 import { goto } from 'src/actions/suite/routerActions';
 import { useDispatch, useSelector } from 'src/hooks/suite';
-import { acquireDevice } from 'src/actions/suite/suiteActions';
+import { acquireDevice } from 'src/actions/suite/deviceThunks';
 import { selectDevicesCount, selectDevice } from 'src/reducers/suite/deviceReducer';
 
 import { DeviceStatus } from './DeviceStatus';
