+++ packages/suite/src/components/suite/modals/PassphraseDuplicate/index.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 import { Button, Image } from '@trezor/components';
 
 import { Translation, Modal } from 'src/components/suite';
-import { authorizeDevice, switchDuplicatedDevice } from 'src/actions/suite/suiteActions';
+import { authorizeDevice, switchDuplicatedDevice } from 'src/actions/suite/deviceThunks';
 import { useDevice, useDispatch } from 'src/hooks/suite';
 import { TrezorDevice } from 'src/types/suite';
